Add separate printer functions
Some systems have upwards growing stacks. In such cases a buffer
overflow would not harm the stored return address as the buffer grows in
the wrong direction.

If a program is to be easily exploited in such cases a seperate function
whose return address can be "pre spoiled" is helpful.
See
security.stackexchange.com/questions/44801/smashing-the-stack-if-it-grows-upwards
for more info.